Fix PR suggestions and add a way for the entity to for end the fight
I found that Lord S1 and S2 are not actually bound to throne 1 and 3.
If one dies, throne 3 will always have the defeated state.
This only causes issues when someone joins the fight when one of mantis lords is dead in the second phase. That's why when the host ends the fight, it forces the fight to end on the client side.